refactor(redis_ops): extract protocol module
Extract RESP protocol formatting functions into dedicated protocol.rs module:
- format_resp_command: Formats Redis commands to RESP protocol
- parse_redis_command: Parses Redis commands with quote/escape handling
- format_command_output: Routes command formatting based on output format

All 29 protocol tests moved with the functions. Only format_command_output
is exported for use by other modules.
